pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quark"]

[project.urls]
Homepage = "https://github.com/QUARK-framework/QUARK"
Issues = "https://github.com/QUARK-framework/QUARK/issues"

[project]
name = "QUARK-framework"
version = "0.1.5"
authors = [
    { name = "Philipp Ross", email = "philipp.ross@bmw.de" },
    { name = "Marvin Erdmann", email = "marvin.erdmann@bmw.de" },
    { name = "Andre Luckow", email = "andre.luckow@bmwgroup.com" },
    { name = "Greshma Shaji", email = "greshma.shaji@bmwgroup.com" },
    { name = "Florian Kiwit", email = "florian.kiwit@bmw.de" },
    { name = "Benjamin Decker", email = "benjamin.decker@bmw.de" },
]
description = "Framework for Quantum Computing Application Benchmarking"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "Apache-2.0"
dependencies = [
    "anytree>=2.12.1",
    "inquirer>=3.4.0",
    "matplotlib>=3.10.1",
    "packaging>=24.2",
    "pyyaml>=6.0.2",
]
